requirements.txt
Django>=4.2
python-dotenv
mysqlclient
Pillow
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-ratelimit
selenium
webdriver-manager
fake-useragent
django-allauth
cryptography
serpapi

##
celery
django_cotton
pytz
bs4
redis
numpy